AOS_Inventory_project/Distributed_Inventory/requirements.txt
grpcio==1.56.2
grpcio-tools==1.56.2
protobuf==4.24.0
